Intelli-Gallery/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-firestore==2.14.0
Flask-Cors==4.0.0